Introduce CastIcon
